requirements.txt
# Core dependencies
numpy>=1.21.0
sounddevice>=0.4.5
python-dotenv>=0.19.0
requests>=2.27.0

# Speech recognition (choose one or both)
faster-whisper>=0.9.0
# OR
# openai-whisper>=20230314

# Optional: for development and testing
pytest>=7.0.0
black>=22.0.0
pylint>=2.12.0
